fix: split quality setter method to comply with SRP
The setQuality method on GildedRose had too much responsibilities
(increase, decrease and set to zero).
It was problematic with backstage items returning to zero
when it was to 50 already.
I think it is easier now to check the boundaries and the logic.
Sad that I had to add some comments though... or not...

Mutation "changed conditional boundary" was not being killed with
the bounds '<= 0' and '>= 50' so they have been replaced with
respectively '<= -1' and '>= 51'.
The test cases was added to kill the mutations.

Refactored GildedRoseParameterizedNestedUnitTest
and removed 'assertStateTransition' method
because it was adding useless complexity (for reading).